Story:  IOC - Refactor SpiderHome. [id=85875]
git-svn-id: https://geekscape.org/svn/boost/trunk@2920 05db6a09-9903-0410-bedf-ba9bc4063e48

+++ src/test/java/au/net/netstorm/boost/test/specific/TimePointProvider.java
@@ -13,7 +13,7 @@ public final class TimePointProvider implements DataProvider {
 
     public Data get() {
         // FIX 64166 Range support required here.
-        long time = ((Long) random.provide(Long.class)).longValue();
+        long time = random.provide(Long.class);
         long positive = time < 0 ? -time : time;
         return new DefaultTimePoint(positive);
     }
